heuristics: pv move sorting
This commit adds a new heuristic: being able to follow the PV line to
reduce nodes searched. This means that we will prioritize the PV line as
that's "the best" line so far.

+++ src/evaluation/perft.h
@@ -38,7 +38,7 @@ private:
     constexpr static inline void search(const BitBoard& board, uint8_t depth, bool printMove = false)
     {
         const auto allMoves = engine::getAllMoves(board);
-        uint16_t legalMoves = 0;
+        uint64_t legalMoves = 0;
 
         if (depth == 0) {
             for (const auto& move : allMoves.getMoves()) {
